reorganized testing folders and added more snapshots
