+++ src/__tests__/combination.test.js
@@ -2,8 +2,8 @@ import React from 'react'
 import { render, cleanup } from 'react-testing-library'
 
 import createMockRaf from 'mock-raf'
-import { ColorAndHeight } from '../stories/tests/combination'
-import { Globals } from '../src/targets/web'
+import { ColorAndHeight } from '../../stories/tests/combination'
+import { Globals } from '../targets/web'
 
 test('width transform', () => {
   const mockRaf = createMockRaf()
@@ -21,17 +21,23 @@ test('width transform', () => {
   mockRaf.step({ count: 10 })
 
   expect(box.style.backgroundColor).toBe('rgb(144, 56, 144)')
-  expect(parseFloat(box.style.height.slice(0, -2))).toBeCloseTo(325.6051453648133)
+  expect(parseFloat(box.style.height.slice(0, -2))).toBeCloseTo(
+    325.6051453648133
+  )
 
   mockRaf.step({ count: 10 })
 
   expect(box.style.backgroundColor).toBe('rgb(231, 12, 231)')
-  expect(parseFloat(box.style.height.slice(0, -2))).toBeCloseTo(462.6970463957336)
+  expect(parseFloat(box.style.height.slice(0, -2))).toBeCloseTo(
+    462.6970463957336
+  )
 
   mockRaf.step({ count: 10 })
 
   expect(box.style.backgroundColor).toBe('rgb(251, 2, 251)')
-  expect(parseFloat(box.style.height.slice(0, -2))).toBeCloseTo(493.36314302940417)
+  expect(parseFloat(box.style.height.slice(0, -2))).toBeCloseTo(
+    493.36314302940417
+  )
 
   mockRaf.step({ count: 100 })
