+++ src/animated/__tests__/Animated.test.js
@@ -1,0 +1,13 @@
+import Animated from './../Animated'
+
+let TestedClass = new Animated()
+
+describe('Animated', () => {
+  beforeEach(() => {
+    TestedClass = new Animated()
+  })
+
+  it('instantiates', () => {
+    expect(TestedClass).toBeTruthy()
+  })
+})
